+++ sdk/azidentity/default_azure_credential.go
@@ -159,6 +159,7 @@ func NewDefaultAzureCredential(options *DefaultAzureCredentialOptions) (*Default
 			ClientOptions:              options.ClientOptions,
 			DisableInstanceDiscovery:   options.DisableInstanceDiscovery,
 			TenantID:                   options.TenantID,
+			disableIdentityBindingMode: true,
 		})
 		if err == nil {
 			creds = append(creds, wic)