+++ sdk/azidentity/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Breaking Changes
 
+> These changes affect only code written against a beta version such as v1.13.0-beta.1
+- Removed identity binding mode support from `DefaultAzureCredential`. To use this feature, use
+  `WorkloadIdentityCredential` directly instead.
+
 ### Bugs Fixed
 
 ### Other Changes